Modular Structure for Runner and Config
