+++ lib/local.js
@@ -5,10 +5,9 @@ var Log = require('./logger'),
   https = require('https'),
   windows = ((process.platform.match(/win32/) || process.platform.match(/win64/)) !== null),
   localBinary = __dirname + '/BrowserStackLocal' + (windows ? '.exe' : ''),
-  utils = require('./utils'),
-  config = require('./config');
+  utils = require('./utils');
 
-var Tunnel = function Tunnel(key, port, uniqueIdentifier, callback) {
+var Tunnel = function Tunnel(key, port, uniqueIdentifier, config, callback) {
   var that = {};
 
   function tunnelLauncher() {
@@ -36,7 +35,8 @@ var Tunnel = function Tunnel(key, port, uniqueIdentifier, callback) {
 
     setTimeout(function() {
       if (!running) {
-        utils.alertBrowserStack('Tunnel launch timeout', 'Stdout:\n' + data);
+        logger.error('BrowserStackLocal failed to launch within 30 seconds.');
+        process.exit(1);
       }
     }, 30 * 1000);
