+++ lib/server.js
@@ -7,7 +7,6 @@ var Log = require('./logger'),
   fs = require('fs'),
   qs = require('querystring'),
   utils = require('./utils'),
-  config = require('../lib/config'),
   proxyServer = require('./proxy').proxyServer,
   chalk = require('chalk'),
   mime = require('mime'),
@@ -15,7 +14,7 @@ var Log = require('./logger'),
   vm = require('vm');
 
 
-exports.Server = function Server(bsClient, workers) {
+exports.Server = function Server(bsClient, workers, config) {
 
   var testFilePaths = (Array.isArray(config.test_path) ? config.test_path : [ config.test_path ])
     .map(function (path) {
