+++ lib/utils.js
@@ -50,83 +50,7 @@ var objectSize = function objectSize(obj) {
   return size;
 };
 
-var alertBrowserStack = function alertBrowserStack(subject, content, params, fn) {
-  if (!config) {
-    config = require('./config');
-  }
-
-  var endpoint = config.alert_endpoint || 'http://www.browserstack.com/automate/alert';
-  var urlObject = url.parse(endpoint);
-
-  var context = config.alert_context || 'Runner alert';
-  logger.info('[%s] [%s] %s, %s', new Date().toISOString(), context, subject, content);
-
-  if (typeof fn !== 'function') {
-    if (typeof params === 'function') {
-    } else {
-      fn = function() {
-        process.exit('SIGINT');
-      };
-    }
-  }
-
-  if (!params || typeof(params) !== 'object') {
-    params = {};
-  }
-
-  params.subject = subject;
-  params.content = content;
-  params.context = context;
-
-  var body = querystring.stringify(params);
-  var options = {
-    hostname: urlObject.hostname,
-    port: urlObject.port,
-    path: urlObject.path,
-    method: 'POST',
-    auth: config.username + ':' + config.key,
-    headers: {
-      'Content-Length': body.length
-    }
-  };
-
-  var callback = function(res) {
-    var response = '';
-    res.setEncoding('utf8');
-    res.on('data', function(chunk) {
-      response += chunk;
-    });
-    res.on('end', function() {
-      if (res.statusCode !== 200) {
-        var message;
-        if (res.headers['content-type'].indexOf('json') !== -1) {
-          var resp = JSON.parse(response);
-          message = resp.message;
-          message += ' - ' + resp.errors.map(function(err) {
-            return '`' + err.field + '`' + ' ' + err.code;
-          }).join(', ');
-        } else {
-          message = response;
-        }
-        if (!message && res.statusCode === 403) {
-          message = 'Forbidden';
-        }
-        fn(new Error(message));
-      } else {
-        fn(null, JSON.parse(response));
-      }
-    });
-  };
-
-  var request = http.request(options, callback);
-  request.write(body);
-  request.end();
-
-  return request;
-};
-
 exports.titleCase = titleCase;
 exports.uuid = uuid;
 exports.browserString = browserString;
 exports.objectSize = objectSize;
-exports.alertBrowserStack = alertBrowserStack;
