+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "browserstack-runner",
   "description": "A command line interface to run browser tests over BrowserStack",
-  "version": "0.4.4",
+  "version": "0.4.5",
   "homepage": "https://github.com/browserstack/browserstack-runner",
   "repository": {
     "type": "git",
@@ -28,6 +28,7 @@
   "bin": {
     "browserstack-runner": "bin/cli.js"
   },
+  "main": "bin/runner.js",
   "scripts": {
     "lint": "node_modules/.bin/jshint lib/*.js bin/ tests/*.js",
     "test-unit": "node_modules/.bin/mocha tests/unit",
